==> 텍스트마이닝 활용 기준금리 예측/4_TrainingSet 생성.ipynb <==
# 1. Training Set 생성

import pandas as pd
import numpy as np


df = pd.DataFrame()
callrate = pd.read_csv('./data/market/callrate_lable.csv')

#bond reports
bond = pd.read_csv('./data/bondreports/bondreports_preprocessed.csv')
bond['date'] = '20'+ bond['date']
bond['date'] = pd.to_datetime(bond['date'],format='%Y.%m.%d', errors='ignore')
callrate['date'] = pd.to_datetime(callrate['date'],format='%Y.%m.%d', errors='ignore')

df = pd.merge(bond, callrate, on='date', how='left')
df = df.fillna('')
df['ngrams in doc'] = df["ngram"] + ',' +  df["tokens"] 
df[['date','ngrams in doc','lable']].to_csv('./data/model/training_set.csv', mode='w', index=False)


#minutes
minutes = pd.read_csv('./data/minutes/minutes_preprocessed.csv')
minutes['date'] = pd.to_datetime(minutes['date'],format='%Y%m%d', errors='ignore')
callrate['date'] = pd.to_datetime(callrate['date'],format='%Y.%m.%d', errors='ignore')

df = pd.DataFrame()
df = pd.merge(minutes.loc[minutes.date <= '2017.12.31'], callrate, on='date', how='left')
df = df.fillna('')
df['ngrams in doc'] = df["Foreign Currency tokens"] + ',' +  df["Financial Markets tokens"]  + ',' +  df["Foreign Currency ngram"]  + ',' +  df["Financial Markets ngram"] 
df[['date','ngrams in doc','lable']].to_csv('./data/model/training_set.csv', mode='a', header=False, index=False)



import pandas as pd
from os import listdir
from os.path import isfile, join

#news
source_folder = "./data/news/"
news_ls = [f for f in listdir(source_folder) if isfile(join(source_folder, f))]
for newsfile in news_ls :
    if newsfile in ["news_2005_preprocessed_n.csv","news_2006_preprocessed_n.csv","news_2007_preprocessed_n.csv","news_2008_preprocessed_n.csv","news_2009_preprocessed_n.csv","news_2010_preprocessed_n.csv","news_201_preprocessed_n.csv"] :
        news = pd.read_csv(source_folder + newsfile)
        news['date'] = pd.to_datetime(news['date'],format='%Y.%m.%d', errors='ignore')
        callrate['date'] = pd.to_datetime(callrate['date'],format='%Y.%m.%d', errors='ignore')

        df = pd.DataFrame()
        df = pd.merge(news, callrate, on='date', how='left')
        df = df.fillna('')
        df['ngrams in doc'] = df["tokens"] + ',' +  df["ngram"] 
        df[['date','ngrams in doc','lable']].to_csv('./data/model/training_set.csv', header=False, mode='a', index=False)

traning_set = pd.read_csv('./data/model/training_set.csv')
prev = 0
for i, row in traning_set.sort_values(by=['date'], ascending=False).iterrows() :
    if row['lable'] is None :
        print(i)
    
    prev = row['lable']

traning_set.sort_values(by=['date'], ascending=False).head()

traning_set.loc[traning_set.lable==1.0].shape

traning_set.loc[traning_set.lable==-1.0].shape
