==> 텍스트마이닝 활용 기준금리 예측/06_Naive Bayes Classifier.ipynb <==
# Naive Bayes Classifier 

## 1) Traning set 불러오기

# training set, test set 만들기
import pandas as pd

data_set = pd.read_csv('./data/model/training_set.csv').dropna()
data_set = data_set[(data_set.lable==1.0) | (data_set.lable==-1.0)]

data_set.head()

## 2) Traning

# traning : traning set 으로 토큰별 빈도수 계산
from collections import defaultdict

def set_ngram_list(data_set) :
    # 토큰별로 문서내 빈도수 카운팅
    wordfreq = defaultdict(lambda : [0, 0])
    for i, row in data_set.iterrows():
        words = row['ngrams in doc'].replace('@@@',',').replace('.',',').split(',')

        # 문자열 길이가 1인것은 제외
        if len(words[0].split('/')[0]) == 1 :
            continue

        for word in words :
            if row['lable'] == 1 :
                wordfreq[word][1] += 1
            elif row['lable'] == -1 :
                wordfreq[word][0] += 1

        if i % 10000 == 0 :
            print("{} / {}".format(i, data_set.shape[0]))
            #break

    pd.DataFrame.from_dict(wordfreq, orient='index').to_csv('./data/model/ngram_list.csv')
    
set_ngram_list(data_set)

df = pd.read_csv('./data/model/ngram_list.csv')

df.shape

import math

# 예측 및 모델 검증
def predict(test_set, dic_df, sum_1, sum_0, intensity) :
    dic_df["polarity score"] = dic_df["w|pos"]/dic_df["w|neg"]
    dic_df["intensity"] = [x if x > 1 else 1/x for x in dic_df['polarity score']]
    dic_df = dic_df.loc[dic_df.intensity>intensity]

    dic = dic_df.to_dict('index')
    
    prob_pos_tot = math.log(sum_1/(sum_1+sum_0))
    prob_neg_tot = math.log(sum_0/(sum_1+sum_0))
    
    TP = FN = FP = TN = cnt = 0
    predict=[]

    for i, row in test_set.iterrows():
        ngrams = row['ngrams in doc'].split(',')
        lable = row['lable']
        prob_predict = prob_pos = prob_neg = 0

        tmp_token = ""
        tmp_pos = 0
        tmp_neg = 0
        try :
            for token in ngrams :
                if token == '' :
                    continue 
                
                if token in dic.keys() :
                    prob_pos += dic[token]['log(w|pos)']
                    prob_neg += dic[token]['log(w|neg)']
                    
            prob_pos = math.exp(prob_pos_tot + prob_pos)
            prob_neg = math.exp(prob_neg_tot + prob_neg)
        except Exception as e :
            print("===============")
                        
            print(str(e))
            print("prob_pos_tot : {:10.4f}".format(prob_pos_tot))
            print("prob_pos : {:10.4f}".format(prob_pos))
            print("prob_neg_tot : {:10.4f}".format(prob_neg_tot))
            print("prob_neg : {:10.4f}".format(prob_neg))
            print("sum_1 : {:10.4f}".format(sum_1))
            print("sum_0 : {:10.4f}".format(sum_0))
            print(ngrams)


        if prob_pos+prob_neg == 0 :
            prob_predict = 0
        else :
            prob_predict = prob_pos/(prob_pos+prob_neg)

        if prob_predict > 0.5 and lable == '1' :
            predict.append(1.0)
            TP += 1
        elif prob_predict > 0.5 and lable == '-1' :
            predict.append(1.0)
            FP += 1
        elif prob_predict < 0.5 and lable == '1' :
            predict.append(-1.0)
            FN += 1
        elif prob_predict < 0.5 and lable == '-1' :
            predict.append(-1.0)
            TN += 1
            
    print('TP={} / TN={} / FP={} / FN={}'.format(TP, TN, FP, FN))
    #ACC = (TP + TN)  / (전체 데이타 수 = P + N)
    ACC = (TP + TN) / (TP + TN + FP + FN) * 100
    # ERR = (FN+FP) / (전체 데이타수 = P+N)
    ERR = (FN+FP) / (TP + TN + FP + FN) * 100
    # SN = (TP) / P
    SN = (TP) / (TP + FN) * 100
    # PREC = TP / (TP+FP)
    PREC = TP / (TP+FP) * 100
    print("ACC={:7.3f}% / ERR = {:7.3f}% / SN = {:7.3f}% / PREC = {:7.3f}%".format(ACC, ERR, SN, PREC))

#dictionary 파일 저장
import pandas as pd
import numpy as np
from sklearn.utils import shuffle


def traning(data_set) :
    intensity = 1.3
    no_bagging = 1
    k = 0.5
    
    df = pd.read_csv('./data/model/ngram_list.csv', index_col="Unnamed: 0")
    df["freq"] =0
    df["w|pos"] =0
    df["w|neg"] =0

    for i in range(0,no_bagging) :
        print("========== bagging : " + str(i+1))
        data_set = shuffle(data_set)
        training_set = data_set[:int(len(data_set)*0.9)]
        test_set = data_set[int(len(data_set)*0.9):]
        
        sum_1 = df['1'].sum()#training_set.loc[training_set['lable']>0].shape[0]
        sum_0 = df['0'].sum()#training_set.loc[training_set['lable']<0].shape[0]

        df["w|pos"] += (df['1']+k)/(sum_1+k*2)
        df["w|neg"] += (df['0']+k)/(sum_0+k*2)
        df["log(w|pos)"] = np.log(df["w|pos"])
        df["log(w|neg)"] = np.log(df["w|neg"])
        
        predict(test_set, df, sum_1, sum_0, intensity)
 
    df["freq"] = df['0']+df['1'] # 전체 토큰별 빈도수 계산
    #df = df.loc[df.freq > 15] # 빈도가 15 초과인것만 필터링  
    df["w|pos"] = df["w|pos"]/no_bagging
    df["w|neg"] = df["w|neg"]/no_bagging        
    #df["log(w|pos)"] = np.log(df["w|pos"])
    #df["log(w|neg)"] = np.log(df["w|neg"])
    df["polarity score"] = df["w|pos"]/df["w|neg"]
    df["intensity"] = [x if x > 1 else 1/x for x in df['polarity score']]
    df["lable"] = [1 if x > 1 else -1 for x in df['polarity score']]
    df.loc[df.intensity>intensity].to_csv('./data/model/bok_dictionary.csv')
    
traning(data_set)

import pandas as pd
ff = pd.read_csv('./data/model/bok_dictionary.csv', index_col="Unnamed: 0")
tmp = ff[ff['lable']==1].sort_values('log(w|pos)', ascending=False )
tmp[tmp.index.str.contains(";")].head(30)

tmp[tmp.index=='경제/NNG;성장/NNG']

tmp = ff[ff['lable']==-1].sort_values('log(w|neg)', ascending=False )
tmp[tmp.index.str.contains(";")].head(30)

## 3) prediction & 검증

def r_squared(y_true, y_hat):
    ssr = 0
    sst = 0
    e = np.subtract(y_true, y_hat)
    y_mean = np.mean(y_true)
    for item in e:
        ssr += item**2
    for item in y_true:
        sst += (item - y_mean)**2
    r2 = 1 - ssr / sst
    return r2

#r_squared(lable_ls, predict)
