==> 텍스트마이닝 활용 기준금리 예측/07_MarketApproach.ipynb <==
# Market Approach 측정

# 의사록 읽기
import pandas as pd
minutes_ls = pd.read_csv("./data/minutes/minutes_preprocessed.csv").fillna('')

df = pd.read_csv('./data/model/bok_dictionary.csv', index_col='Unnamed: 0')
dic = df.to_dict('index')

index = 1
minutes_ls['Foreign Currency'][index]+'@@@'+minutes_ls['Financial Markets'][index]

import numpy as np
import pandas as pd

df = pd.DataFrame(columns=['date','tone','no_hawkish_tone_s', 'no_dovish_tone_s', 'tones', 'no_hawkish_features', 'no_dovish_features', 'hawkish_features', 'dovish_features' ])

for i, minutes in minutes_ls.iterrows() :
    hawkish_features_ls = []
    dovish_features_ls = []
    no_hawkish_features_ls = []
    no_dovish_features_ls = []
    tone_s_ls = []
    no_hawkish_tone_s = no_dovish_tone_s = tone = 0

    sentences = minutes["ngram"].split('@@@')
    
    for sentence in sentences :
        no_hawkish_features = no_dovish_features = tone_s = 0
        hawkish_features = []
        dovish_features = []
        for token in sentence.split(',') :
            if token in dic.keys() :
                if dic[token]['lable'] == 1.0 :
                    no_hawkish_features += 1
                    hawkish_features.append(token)
                if dic[token]['lable'] == -1.0 :
                    no_dovish_features += 1
                    dovish_features.append(token)
                    
        hawkish_features_ls.append(','.join(hawkish_features))
        dovish_features_ls.append(','.join(dovish_features))

        if no_hawkish_features+no_dovish_features > 0 :
            tone_s = (no_hawkish_features-no_dovish_features)/(no_hawkish_features+no_dovish_features)
    
        no_hawkish_features_ls.append(no_hawkish_features)
        no_dovish_features_ls.append(no_dovish_features)
        tone_s_ls.append(tone_s)
        
        no_hawkish_tone_s += 1 if tone_s > 0 else 0
        no_dovish_tone_s += 1 if tone_s < 0 else 0
    
    if no_hawkish_tone_s+no_dovish_tone_s > 0 :
        tone = (no_hawkish_tone_s-no_dovish_tone_s)/(no_hawkish_tone_s+no_dovish_tone_s)
    
    #print(no_hawkish_features_ls)
    #print(no_dovish_features_ls)
    #print(tone_s_ls)
    
    df.loc[len(df)] = [minutes['date']
                       ,tone,no_hawkish_tone_s,no_dovish_tone_s
                       ,','.join(str(x) for x in tone_s_ls)
                       ,','.join(str(x) for x in no_hawkish_features_ls)
                       ,','.join(str(x) for x in no_dovish_features_ls)
                       ,'@@@'.join(hawkish_features_ls)
                       ,'@@@'.join(dovish_features_ls)]

df.to_csv('./results/market_approach.csv')

df[['date','tone']].to_csv('./results/market_approach_index.csv', index=False)

for x, y in zip(df['hawkish_features'], df['dovish_features']) :
    print(x.split('@@@')[1])
    print(y.split('@@@')[1])
    break

dic['완만/NNG;속도/NNG;확장/NNG']

df

import pandas as pd
from os import listdir
from os.path import isfile, join

#news
source_folder = "./data/news/"

news = pd.read_csv(source_folder + "news_2005_preprocessed_n.csv")


news.head()
